+++ frontend/src/charts/rateBarChart/constants.ts
@@ -1,0 +1,26 @@
+const Y_AXIS_LABEL_HEIGHT = 20
+const BAR_PADDING = 0.2
+const LABEL_SWAP_CUTOFF_PERCENT = 66
+const MAX_LABEL_WIDTH_BIG = 100
+const MAX_LABEL_WIDTH_SMALL = 50
+const NORMAL_MARGIN_HEIGHT = 20
+const MARGIN = {
+  top: NORMAL_MARGIN_HEIGHT,
+  right: NORMAL_MARGIN_HEIGHT,
+  bottom: NORMAL_MARGIN_HEIGHT + 30,
+  left: NORMAL_MARGIN_HEIGHT,
+}
+const BAR_HEIGHT = 70
+const EXTRA_SPACE_AFTER_ALL = 10
+
+export {
+  BAR_HEIGHT,
+  BAR_PADDING,
+  EXTRA_SPACE_AFTER_ALL,
+  LABEL_SWAP_CUTOFF_PERCENT,
+  MARGIN,
+  MAX_LABEL_WIDTH_BIG,
+  MAX_LABEL_WIDTH_SMALL,
+  NORMAL_MARGIN_HEIGHT,
+  Y_AXIS_LABEL_HEIGHT,
+}
